+++ src/mesh/Triangle.hpp
@@ -108,6 +108,9 @@ public:
   template <typename VECTOR_T>
   void setNormal(const VECTOR_T &normal);
 
+  /// Computes and sets the normal of the triangle, returns the area-weighted normal.
+  const Eigen::VectorXd computeNormal(bool flip = false);
+
   /// Returns a among triangles globally unique ID.
   int getID() const;