+++ src/precice/SolverInterface.hpp
@@ -192,7 +192,7 @@ public:
    * @note
    * The user should call finalize() after this function returns false.
    */
-  bool isCouplingOngoing();
+  bool isCouplingOngoing() const;
 
   /**
    * @brief Checks if new data to be read is available.
@@ -212,7 +212,7 @@ public:
    * This is not recommended due to performance reasons.
    * Use this function to prevent unnecessary reads.
    */
-  bool isReadDataAvailable();
+  bool isReadDataAvailable() const;
 
   /**
    * @brief Checks if new data has to be written before calling advance().
@@ -232,7 +232,7 @@ public:
    * This is not recommended due to performance reasons.
    * Use this function to prevent unnecessary writes.
    */
-  bool isWriteDataRequired ( double computedTimestepLength );
+  bool isWriteDataRequired ( double computedTimestepLength ) const;
 
   /**
    * @brief Checks if the current coupling timestep is completed.
@@ -246,7 +246,7 @@ public:
    *
    * @pre initialize() has been called successfully.
    */
-  bool isTimestepComplete();
+  bool isTimestepComplete() const;
 
   /**
    * @brief Returns whether the solver has to evaluate the surrogate model representation.
@@ -261,7 +261,7 @@ public:
    *
    * @see hasToEvaluateFineModel()
    */
-  bool hasToEvaluateSurrogateModel();
+  bool hasToEvaluateSurrogateModel() const;
 
   /**
    * @brief Checks if the solver has to evaluate the fine model representation.
@@ -276,7 +276,7 @@ public:
    *
    * @see hasToEvaluateSurrogateModel()
    */
-  bool hasToEvaluateFineModel();
+  bool hasToEvaluateFineModel() const;
 
   ///@}
 
@@ -298,7 +298,7 @@ public:
    * @see fulfilledAction()
    * @see cplscheme::constants
    */
-  bool isActionRequired ( const std::string& action );
+  bool isActionRequired ( const std::string& action ) const;
 
   /**
    * @brief Indicates preCICE that a required action has been fulfilled by a solver.
@@ -341,14 +341,14 @@ public:
    * @param[in] meshName the name of the mesh
    * @returns the id of the corresponding mesh
    */
-  int getMeshID ( const std::string& meshName );
+  int getMeshID ( const std::string& meshName ) const;
 
   /**
    * @brief Returns a id-set of all used meshes by this participant.
    *
    * @returns the set of ids.
    */
-  std::set<int> getMeshIDs();
+  std::set<int> getMeshIDs() const;
 
   /**
    * @brief Returns a handle to a created mesh.
@@ -382,7 +382,7 @@ public:
    * @param[in] meshID the id of the mesh
    * @returns the amount of the vertices of the mesh
    */
-  int getMeshVertexSize(int meshID);
+  int getMeshVertexSize(int meshID) const;
 
   /**
    * @brief Creates multiple mesh vertices
@@ -426,7 +426,7 @@ public:
     int        meshID,
     int        size,
     const int* ids,
-    double*    positions );
+    double*    positions ) const;
 
   /**
    * @brief Gets mesh vertex IDs from positions.
@@ -447,7 +447,7 @@ public:
     int           meshID,
     int           size,
     const double* positions,
-    int*          ids );
+    int*          ids ) const;
 
   /**
    * @brief Sets mesh edge from vertex IDs, returns edge ID.
@@ -570,7 +570,7 @@ public:
    *
    * @returns the id of the corresponding data
    */
-  int getDataID ( const std::string& dataName, int meshID );
+  int getDataID ( const std::string& dataName, int meshID ) const;
 
   /**
    * @brief Computes and maps all read data mapped to the mesh with given ID.
@@ -712,7 +712,7 @@ public:
     int        dataID,
     int        size,
     const int* valueIndices,
-    double*    values );
+    double*    values ) const;
 
   /**
    * @brief Reads vector data form a vertex
@@ -737,7 +737,7 @@ public:
   void readVectorData (
     int     dataID,
     int     valueIndex,
-    double* value );
+    double* value ) const;
 
   /**
    * @brief Reads scalar data as a block.
@@ -763,7 +763,7 @@ public:
     int        dataID,
     int        size,
     const int* valueIndices,
-    double*    values );
+    double*    values ) const;
 
   /**
    * @brief Reads scalar data of a vertex.
@@ -783,7 +783,7 @@ public:
   void readScalarData (
     int     dataID,
     int     valueIndex,
-    double& value );
+    double& value ) const;
 
   ///@}